+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## [0.4.0] - unreleased
+### Added
+- Allowing dependencies to Interfaces only is now possible
+    - Using the `AllowInterfaces` Validation wrapper
+    - Using the `mustNotDependOn` method on a component
+    - Using the `mustNotDirectlyDependOn` method on an architecture
+
 ## [0.3.1] - 2018-12-12
 ### Fixed
 - Not loading tests into production autoloader anymore